UDIDRESEARCH/backend/fact_checker.py
# Fact-checker module
import requests
import os
from dotenv import load_dotenv

load_dotenv()
SERPAPI_KEY = os.getenv("SERPAPI_KEY")
BING_SEARCH_API_KEY = os.getenv("BING_SEARCH_API_KEY")

class FactChecker:
    def __init__(self):
        pass
    def verify_claim(self, claim):
        # Placeholder: Implement using SerpAPI or Bing Search API
        return {"claim": claim, "verified": False, "source": None}

# ...existing code...
