UDIDRESEARCH/backend/requirements.txt
fastapi
python-dotenv
requests
fpdf
